# Use optimised ByteString to Vector conversion