--- a/stack.yaml
+++ b/stack.yaml
@@ -7,7 +7,7 @@ extra-deps:
 - bits-extra-0.0.1.1
 - hedgehog-0.5.3
 - hw-hspec-hedgehog-0.1.0.4
-- hw-prim-0.6.0.0
+- hw-prim-0.6.2.0
 - hw-rankselect-0.12.0.2
 - hw-rankselect-base-0.3.2.0
 - lazy-csv-0.5.1