### When selecting no columns
Before:

```
$ cat ~/arbornetwrks-small.csv | pv -a | time hw-sv query-lazy --source - --delimiter , --target - --out-delimiter '|' > a.txt
[ 275MiB/s]
9.16s user 1.25s system 170% cpu 6.097 total
```

After:
```
$ cat ~/arbornetwrks-small.csv | pv -a | time hw-sv query-lazy --source - --delimiter , --target - --out-delimiter '|' > b.txt
[ 376MiB/s]
7.18s user 1.19s system 187% cpu 4.473 total
```

### When selecting two columns
Before:
```
$ cat ~/arbornetwrks-small.csv | pv -a | time hw-sv query-lazy --source - --column 0 --column 1 --delimiter , --target - --out-delimiter '|' > a.txt
[ 121MiB/s]
23.15s user 3.35s system 192% cpu 13.757 total
```

After:
```
$ cat sample.csv | pv -a | time hw-sv query-lazy --source - --column 0 --column 1 --delimiter , --target - --out-delimiter '|' > b.txt
[ 147MiB/s]
19.63s user 3.03s system 198% cpu 11.408 total
```